refactor(1): multifile environments - changes after cr #3
