+++ src/services/rules-builder/RulesBuilderTypes.ts
@@ -1,0 +1,5 @@
+export interface RulesContent {
+  markdown: string;
+  label: string;
+  fileName: `${string}.mdc`;
+}
